% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{addSlash}
\alias{addSlash}
\title{Adds Slash to path name}
\usage{
addSlash(targetPath)
}
\arguments{
\item{targetPath}{Path name}
}
\value{
A path name with trailing slash
}
\description{
adds a slash to the end of a path name
}
