% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{convMerraToBin}
\alias{convMerraToBin}
\title{conversion of merra files to binary files}
\usage{
convMerraToBin(in_path, out_path, date_seq, param, multI, silent = TRUE)
}
\arguments{
\item{in_path}{Path of merra files}

\item{out_path}{path where output binary files should be located}

\item{date_seq}{sequence of dates which should be converted}

\item{param}{Parameter in MERRA file that should be converted}

\item{multI}{multiplicator for variable (to be used if numbers after comma are relevant)}

\item{silent}{If true, no detailed output on handling of files is printed to console}
}
\description{
converts merra files to a fast readable binary format
}

