% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{createTestDataSet}
\alias{createTestDataSet}
\title{Creates a MERRA Test Data Set}
\usage{
createTestDataSet(dir, date_seq)
}
\arguments{
\item{dir}{Directory to save the test data set}

\item{date_seq}{length of test data set}
}
\value{
Does not return anything. Writes, as side-effect, two files in the given directory: test_data.feater and MERRA_Test.Rdata, which contain the same information
}
\description{
Creates the test-data set also available as data in the package
}
