% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{FindVars}
\alias{FindVars}
\title{Helper function to find corresponding dataset for given variables}
\usage{
FindVars(varlist,dim,avg,varn)
}
\arguments{
\item{varlist}{data frame of datasets and contained variables}

\item{dim}{requested dataset dimension. 2d or 3d, however only 2d support tested.}

\item{avg}{if true, time averaged variables will be prioritised}

\item{varn}{name of the requested variable}
}
\value{
name of the corresponding dataset
}
\description{
Searches for dataset containing given variables
}
