% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{getFileMerraFinal}
\alias{getFileMerraFinal}
\title{Get Merra Files from Server}
\usage{
getFileMerraFinal(i, fileNameIn, fileNameOut, user, pass)
}
\arguments{
\item{i}{index in list of filenames}

\item{fileNameIn}{list of filenames to download}

\item{fileNameOut}{list of filenames to save results to}

\item{user}{username for geo disc server}

\item{pass}{password for geo disc server}
}
\description{
Function to download a single MERRA file with authentication from gesdisc. This is a helper function to make sParApply work
}
