% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{getMERRADataBox}
\alias{getMERRADataBox}
\title{Get MERRA Files for a given spatial box}
\usage{
getMERRADataBox(lon1, lat1, lon2, lat2, period, params, user, pass,
  runParallel = TRUE)
}
\arguments{
\item{period}{A vector with all dates to download (daily entries in the vector).}

\item{params}{The respective parameter to be downloaded}

\item{user, pass}{username and password for the downloadserver get one at: https://urs.earthdata.nasa.gov/home}

\item{runParallel}{if true, download uses available cores to parallelize downloads}

\item{lon, lat}{center of box to download from. data are downloaded from box lon-0.5,lat-0.625,lon+0.5,lat+0.625}
}
\value{
Does not return anything, but writes, as a side-effect, the downloaded files to disk.
}
\description{
Function to download data from the merra2 dataset
}
