% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\docType{class}
\name{MERRABin-class}
\alias{MERRABin-class}
\alias{MERRABin}
\title{Reference class for Binary gridded timeseries data}
\description{
Reference class for Binary gridded timeseries data
}
\section{Fields}{

\describe{
\item{\code{file}}{path to the binary files}

\item{\code{endian}}{Integer endian used on machine (small or big)}

\item{\code{con}}{The connection to the meta file, once opened}

\item{\code{byteSize}}{The size of the integer values in bytes}

\item{\code{dimX}}{X dimension of the coordinate system}

\item{\code{dimY}}{Y dimension of the coordinate system}

\item{\code{timeTim}}{time dimension of the dataset}

\item{\code{lons,lats}}{the longitude and latitude values in the dataset}

\item{\code{time}}{time values in the dataset, in seconds since 1980-01-01}

\item{\code{grid}}{the longitude/latitude grid}

\item{\code{years}}{The years in the dataset}

\item{\code{counts}}{the number of days per year}

\item{\code{timePerFile}}{The number of hours per day}

\item{\code{sizePerYear}}{The number of bytes saved per year per location}
}}


