% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{MERRABin_getAggregatedCapacitiesTimeSeries}
\alias{MERRABin_getAggregatedCapacitiesTimeSeries}
\title{Returns a timeseries with aggregated capacities per merra point
timeseries is extended to the timeseries covered by this MERRAbin object}
\arguments{
\item{dat}{dataframe which contains at least a column lon, a column lat, a column date and a column cap}
}
\value{
Returns the same dateframe, but capacities are aggregated by merra point and the timeseries is extended to the timeseries of this MERRAbin object
}
\description{
Returns a timeseries with aggregated capacities per merra point
timeseries is extended to the timeseries covered by this MERRAbin object
}

