% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{MERRABin_getClosestTS}
\alias{MERRABin_getClosestTS}
\title{Gets the timeseries from the binary file which is closest to the given longitude/latitude}
\arguments{
\item{lon, lat}{Longitude and Latitude where the closest existing point is found}
}
\value{
The complete timeseries as vector
}
\description{
Gets the timeseries from the binary file which is closest to the given longitude/latitude
}
