% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{MERRABin_getCorrespondingLonLat}
\alias{MERRABin_getCorrespondingLonLat}
\title{Finds the closest MERRA points for all given Lons/Lats}
\arguments{
\item{dat}{dataframe which contains at least a column lon and a column lat}
}
\value{
Returns dat with added column MLon and MLat, which indicates the associated MERRA Point
}
\description{
Finds the closest MERRA points for all given Lons/Lats
}

