% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{MERRABin_getMERRAPoint}
\alias{MERRABin_getMERRAPoint}
\title{Returns the closest MERRA Point}
\arguments{
\item{lon}{Longitude of point to assess}

\item{lat}{Latitude of point to assess}
}
\value{
A vector where the first element is the longitude and the second element is the latitude of the closest MERRA point
}
\description{
returns the coordinates of the MERRA point which is closest to the given point
}

