% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{MERRABin_getShape}
\alias{MERRABin_getShape}
\title{gets the set of timeseries which are related to the MERRA points
in the given shape file}
\arguments{
\item{shapefile}{The shapefile to read}

\item{years}{The years to read}
}
\value{
A tibble of timeseries
}
\description{
gets the set of timeseries which are related to the MERRA points
in the given shape file
}

