% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{MERRABin_getTimeYears}
\alias{MERRABin_getTimeYears}
\title{gets the timeseries of timestamps with limitation to certain years}
\value{
The timestamp timeseries
}
\description{
returns a timeseries of timestamps in UTC
}

