% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{printl}
\alias{printl}
\title{Prints to element to console}
\usage{
printl(a, b)
}
\arguments{
\item{a}{Element A}

\item{b}{Element B}
}
\description{
Prints two elements to the console, concatenating them
}
