% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{readBinMult}
\alias{readBinMult}
\title{Reads multiple bytes from a binary file}
\usage{
readBinMult(con, what, size, elements, endian)
}
\arguments{
\item{con}{Binary connection to file}

\item{what}{Type of variable to read (e.g. Integer, Double, etc.)}

\item{size}{Size of variable in bytes}

\item{elements}{Number of elements}

\item{endian}{Endian type of number (Big or Small)}
}
\value{
The values read in from the binary connection
}
\description{
Reads multiple bytes from a binary file
}
