% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{readParamMerra}
\alias{readParamMerra}
\title{reads merra file from disk
interpolates the data from the neighbouring 4 points with respect to the point given in lon lat}
\usage{
readParamMerra(ncname, pname, lon, lat)
}
\arguments{
\item{ncname}{is the name of the MERRA file}

\item{pname}{is the name of the parameter to read}

\item{lon}{Longitude position to read from}

\item{lat}{Latitude position to read from}
}
\value{
a tibble with the interpolated timeseries at the given point. Also, time is added
}
\description{
reads merra file from disk
interpolates the data from the neighbouring 4 points with respect to the point given in lon lat
}
