% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{sum_}
\alias{sum_}
\title{Helper function to sum up to numbers}
\usage{
sum_(a, b)
}
\arguments{
\item{a, b}{numbers to add}
}
\value{
the sum of the two numbers
}
\description{
Adds two numbers
}

