% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{windOutput}
\alias{windOutput}
\title{Title}
\usage{
windOutput(lon, lat, u10m, u50m, v10m, v50m, disph, speeds_in, power_out,
  nmbTurbines)
}
\arguments{
\item{nmbTurbines}{}
}
\description{
Title
}

