% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{writeBinLong}
\alias{writeBinLong}
\title{Writes a very long vector to a binary file}
\usage{
writeBinLong(bytes, connection, size)
}
\arguments{
\item{bytes}{The files to write}

\item{connection}{The connection to write to}

\item{size}{The element size in bytes}
}
\description{
The R function writeBin is limited by 2^31-1 bytes. This function uses writeBin
iteratively to be able to write longer files
}
