% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{writeSingleYearFile}
\alias{writeSingleYearFile}
\title{Helper function to convert one year of MERRA data to a binary file}
\usage{
writeSingleYearFile(year, listOfFiles, date_seq, param, timePerFile, dimX, dimY,
  out_path, multVar, silent = TRUE)
}
\arguments{
\item{year}{Year to convert}

\item{listOfFiles}{list of MERRA files to convert}

\item{date_seq}{vector with sequence of all dates to be converted (including the given year)}

\item{param}{Parameter in MERRA file that should be converted}

\item{timePerFile}{number of hours per file}

\item{dimX}{X dimension of spatial grid}

\item{dimY}{Y dimension of spatial grid}

\item{out_path}{Path to write file to}

\item{multVar}{multiplicator for variable (to be used if numbers after comma are relevant)}

\item{silent}{If true, no detailed output on handling of files is printed to console}
}
\description{
This is a helper function used to converted one year of MERRA data to a binary file
}

