% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{getMERRADataBox}
\alias{getMERRADataBox}
\title{Get MERRA Files for a given spatial box}
\usage{
getMERRADataBox(targetPath = "", lon1, lat1, lon2, lat2, period, params, user,
  pass, runParallel = TRUE, outputofParallelProc = "", dim, avg)
}
\arguments{
\item{targetPath}{Path to save MERRA ncdf files to}

\item{lon1}{Box coordinate}

\item{lat1}{Box coordinate}

\item{lon2}{Box coordinate}

\item{lat2}{Box coordinate}

\item{period}{A vector with all dates to download (daily entries in the vector).}

\item{params}{The respective parameter to be downloaded}

\item{user, pass}{username and password for the downloadserver get one at: https://urs.earthdata.nasa.gov/home}

\item{runParallel}{if true, download uses available cores to parallelize downloads}

\item{outputOfParallelProc}{If different from "", output files are written to the specified directory}
  
\item{dim}{requested dataset dimension. 2d or 3d, however only 2d support tested at the moment.}
  
\item{avg}{if true, time averaged variables will be prioritised}  
}
\value{
Does not return anything, but writes, as a side-effect, the downloaded files to disk.
}
\description{
Function to download data from the merra2 dataset
}
