% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA_functions.R
\name{MERRABin_getClosestTS}
\alias{MERRABin_getClosestTS}
\alias{MERRABin_getClosestTS}
\title{Gets timeseries at lon/lat point}
\arguments{
\item{lon, lat}{Longitude and Latitude where the closest existing point is found}

\item{lon, lat}{Longitude and Latitude where the closest existing point is found}

\item{years}{Years to load from file}
}
\value{
The complete timeseries as vector

The complete timeseries as vector
}
\description{
Gets the timeseries from the binary file which is closest to the given longitude/latitude

Gets the timeseries from the binary file which is closest to the given longitude/latitude
}
